refactor risk management plan into risk controller
